fix: address code review issues for WOP-1465 (WOP-1465)
- Move ModelProvider and related types (ModelClient, ModelQueryOptions,
  ModelResponse, Tool, ToolCall, ToolResult) to src/plugin-types/provider.ts
  so plugin-types no longer imports from core-internal src/types/provider.js
- Re-export moved types from src/types/provider.ts for backward compatibility
- Export ModelProvider from src/plugin-types/index.ts
- Fix unsound getLLMProvider() by using providerRegistry.getProvider(id)?.provider
  instead of the invalid listProviders().find() + as unknown as ModelProvider cast
